requirements.txt
numpy
pillow
pip
attrs
wheel
h5py
pytz
idna
multidict
propcache
setuptools
scipy
threadpoolctl
joblib
pandas
python-dateutil
yarl
aiohttp
aiosignal
frozenlist
aiohappyeyeballs
certifi
scikit-learn
urllib3
six
requests
python-dotenv
aiofiles